### Coppied Send version from Jasper ###
Last evening we did some drastic revisions to the data structure on jaspers laptop due to an authorisation problem he coulnd't push this version to the remote server and send it to me directly.  Because I'm now revealing this revised version the next day and need it up into the git. I copied it locally and will now push it to later be merged with the current feature branch.